feat(automate): auto fill all required fields
By setting the appName, the appVersion and the sources,
the sonar-project.properties file is no longer mandatory.
It will still be needed if you wish to add test coverage but this
could be automated as well for javascript and python if we decide on
a shared pattern for all projects.

BREAKING CHANGES:
This will rename all of the current project in sonarqube and thus reset
their dept monitoring to zero. To avoid that, the projects would have
to be renamed by hand.